+++ docs/source/changelog.rst
@@ -5,7 +5,7 @@ ChangeLog
 0.6.0 (pending)
 ===============
 
-      * Added Cinema2K write support.
+      * Added Cinema2K, Cinema4K write support.
       * Added lxml requirement.
       * added set_printoptions, get_printoptions function
       * dropped support for Python 2.6, added support for Python 3.4